requirements.txt
cachetools==2.1.0
certifi==2018.8.24
chardet==3.0.4
Click==7.0
Flask==1.0.2
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
httplib2==0.11.3
httplib2shim==0.0.3
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
oauth2client==4.1.3
pyasn1==0.4.4
pyasn1-modules==0.2.2
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
rsa==4.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
Werkzeug==0.14.1
wincertstore==0.2
